Update generator command to `kubik:media_library:install`
