+++ lib/generators/kubik/media_library/templates/config/initializers/shrine.rb
